Iterate through matches from right to left
Since we manipulated the string from left to right, there is a risk that the
value of the argument has more chars than the {0}-formatting we're replacing.
Since the matches is done on the template string, this makes further matches
insterted in the wrong place. By instead manipulating the string from right
to left, we solve the problem.